Making OOP-Classes
Creating Object Orientated Classes for Othello.
Game is the driver class.
Board will draw the board.
Position will determine the state of the board position (white, black or empty).
PlayablePosition is whether or not the played move complies with the rule.
UnplayablePosition will contain tiles that cannot be played throughout the duration of the game.
Player will store names of players.