Adding methods
Tring to implement the board as a 1D array.

//TODO
Add board position labelling so that user can input movement coordinates.

Implement method for making a move. It has to be able to determine if the position is playable or not. Also, it has to determine whether to flip other pieces or not.

Make a method to load and save the game. Passing the board as a String parameter.

Implement exceptions handling.