Initialized Board
Successfully initialized board by extending the abstract class Position and initializing the concrete class.
Deleted codes that are not currently useful.
//TODO
Find a way to switch between players so that when p1 makes a move a black piece is played, alternatively when p2 makes a move a white piece is played.